feat: add ES build for tree-shaking
