+++ .gitignore
@@ -52,3 +52,5 @@ jspm_packages
 
 
 # End of https://www.gitignore.io/api/node
+
+pkg
