+++ package.json
@@ -2,9 +2,10 @@
   "name": "li",
   "version": "1.3.0",
   "description": "Parse the Links header format and return a javascript object.",
-  "main": "lib/index.js",
+  "main": "src/index.js",
   "scripts": {
-    "test": "mocha"
+    "test": "mocha",
+    "prepare": "pika-pack"
   },
   "repository": {
     "type": "git",
@@ -23,7 +24,33 @@
   ],
   "license": "MIT",
   "devDependencies": {
-    "mocha": "~1.7.4",
-    "chai": "~1.9.0"
+    "@pika/pack": "^0.5.0",
+    "@pika/plugin-build-node": "^0.9.2",
+    "@pika/plugin-build-umd": "^0.9.2",
+    "@pika/plugin-build-web": "^0.9.2",
+    "@pika/plugin-standard-pkg": "^0.9.2",
+    "chai": "~1.9.0",
+    "mocha": "~1.7.4"
+  },
+  "@pika/pack": {
+    "pipeline": [
+      [
+        "@pika/plugin-standard-pkg",
+        {
+          "exclude": [
+            "__tests__/**/*"
+          ]
+        }
+      ],
+      [
+        "@pika/plugin-build-node"
+      ],
+      [
+        "@pika/plugin-build-web"
+      ],
+      [
+        "@pika/plugin-build-umd"
+      ]
+    ]
   }
 }
